Phase 1 complete: per-line pseudobulk, gene type, example QC report, docs
- stream: stage each plate's gene rows split by cell line, reduce per (plate, line).
  A plate-wide aggregation sat at the 2 GB DuckDB cap and OOMed on the 3-plate build
  (incl. 572e993 on main). Rebuild 9.3 min, peak RSS 2.5 GB; output identical row for
  row to the last good build (56.7M pseudobulk, 67.4M logFC rows).
- genes.parquet: gene is INTEGER, same as pseudobulk; make phase1 builds it if missing
- QC report: dropped-condition count instead of 90 raw ids; rounded library size
- reports/example_qc_phase1.md: 3-plate QC report for the case study
- phase1.md: status complete, §13 rewritten for plates 1-3 with throughput numbers;
  step0_findings: plate triplets, DepMap, pass_filter; roadmap: Phase 1 ticked

+++ scripts/fetch_gene_metadata.py
@@ -44,7 +44,7 @@ def write_genes(con, source: Path, out: Path) -> int:
     con.execute(
         f"""
         COPY (
-            SELECT token_id::BIGINT AS gene, gene_symbol, ensembl_id
+            SELECT token_id::INTEGER AS gene, gene_symbol, ensembl_id
             FROM read_parquet({sql_str(source)}) ORDER BY gene
         ) TO {sql_str(out)} (FORMAT parquet, COMPRESSION zstd)
         """
